+++ src/oci/management_agent/models/__init__.py
@@ -8,12 +8,18 @@ from .availability_history_summary import AvailabilityHistorySummary
 from .create_management_agent_install_key_details import CreateManagementAgentInstallKeyDetails
 from .deploy_plugins_details import DeployPluginsDetails
 from .management_agent import ManagementAgent
+from .management_agent_aggregation import ManagementAgentAggregation
+from .management_agent_aggregation_collection import ManagementAgentAggregationCollection
+from .management_agent_aggregation_dimensions import ManagementAgentAggregationDimensions
 from .management_agent_error import ManagementAgentError
 from .management_agent_image import ManagementAgentImage
 from .management_agent_image_summary import ManagementAgentImageSummary
 from .management_agent_install_key import ManagementAgentInstallKey
 from .management_agent_install_key_summary import ManagementAgentInstallKeySummary
 from .management_agent_plugin import ManagementAgentPlugin
+from .management_agent_plugin_aggregation import ManagementAgentPluginAggregation
+from .management_agent_plugin_aggregation_collection import ManagementAgentPluginAggregationCollection
+from .management_agent_plugin_aggregation_dimensions import ManagementAgentPluginAggregationDimensions
 from .management_agent_plugin_details import ManagementAgentPluginDetails
 from .management_agent_plugin_summary import ManagementAgentPluginSummary
 from .management_agent_summary import ManagementAgentSummary
@@ -24,7 +30,6 @@ from .work_request_error import WorkRequestError
 from .work_request_log_entry import WorkRequestLogEntry
 from .work_request_resource import WorkRequestResource
 from .work_request_summary import WorkRequestSummary
-from .work_submission_key import WorkSubmissionKey
 
 # Maps type names to classes for management_agent services.
 management_agent_type_mapping = {
@@ -32,12 +37,18 @@ management_agent_type_mapping = {
     "CreateManagementAgentInstallKeyDetails": CreateManagementAgentInstallKeyDetails,
     "DeployPluginsDetails": DeployPluginsDetails,
     "ManagementAgent": ManagementAgent,
+    "ManagementAgentAggregation": ManagementAgentAggregation,
+    "ManagementAgentAggregationCollection": ManagementAgentAggregationCollection,
+    "ManagementAgentAggregationDimensions": ManagementAgentAggregationDimensions,
     "ManagementAgentError": ManagementAgentError,
     "ManagementAgentImage": ManagementAgentImage,
     "ManagementAgentImageSummary": ManagementAgentImageSummary,
     "ManagementAgentInstallKey": ManagementAgentInstallKey,
     "ManagementAgentInstallKeySummary": ManagementAgentInstallKeySummary,
     "ManagementAgentPlugin": ManagementAgentPlugin,
+    "ManagementAgentPluginAggregation": ManagementAgentPluginAggregation,
+    "ManagementAgentPluginAggregationCollection": ManagementAgentPluginAggregationCollection,
+    "ManagementAgentPluginAggregationDimensions": ManagementAgentPluginAggregationDimensions,
     "ManagementAgentPluginDetails": ManagementAgentPluginDetails,
     "ManagementAgentPluginSummary": ManagementAgentPluginSummary,
     "ManagementAgentSummary": ManagementAgentSummary,
@@ -47,6 +58,5 @@ management_agent_type_mapping = {
     "WorkRequestError": WorkRequestError,
     "WorkRequestLogEntry": WorkRequestLogEntry,
     "WorkRequestResource": WorkRequestResource,
-    "WorkRequestSummary": WorkRequestSummary,
-    "WorkSubmissionKey": WorkSubmissionKey
+    "WorkRequestSummary": WorkRequestSummary
 }